Split bcl2fastq(2) into two rules. Parallelize bcl2fastq2 per bases mask.
Write out one bcl2fastq2 SampleSheet.csv per bases_mask found in the config.

Allow comma-separated list of sequences for barcode1.
This enables demux to process pools for barcodes, like for a 10x protocol,
without explicitly putting each barcode on new line in the web interface.

Fix number of template reads, fix missing Undetermined for un-indexed runs.

+++ digestiflow_demux/wrappers/bcl2fastq2/wrapper.py
@@ -3,7 +3,12 @@
 This file is part of Digestify Demux.
 """
 
+import glob
+import os
 from snakemake import shell
+
+from digestiflow_demux.bases_mask import return_bases_mask
+from digestiflow_demux.snakemake_support import build_sample_map
 
 __author__ = "Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>"
 
@@ -17,10 +22,14 @@ if barcode_mismatches is None:
 # More than 8 threads will not work for bcl2fastq.
 bcl2fastq_threads = min(8, snakemake.config["cores"])  # noqa
 
-if snakemake.params.read_structure:  # noqa
-    bases_mask = "--use-bases-mask " + snakemake.params.read_structure  # noqa
-else:
-    bases_mask = ""
+# Get bases mask for the current sample sheet
+planned_reads = snakemake.config["flowcell"]["planned_reads"]  # noqa
+bases_mask = os.path.basename(os.path.dirname(snakemake.input.sheet))  # noqa
+bases_mask_illumina = return_bases_mask(planned_reads, bases_mask)
+
+# Get sample map to get sample numbering correct
+sample_map = build_sample_map(snakemake.config["flowcell"])  # noqa
+sample_map["Undetermined"] = "S0"
 
 shell(
     r"""
@@ -36,19 +45,20 @@ trap "rm -rf $TMPDIR" EXIT
 # -------------------------------------------------------------------------------------------------
 # Print Sample Sheet
 
-head -n 10000 {snakemake.params.output_dir}/SampleSheet.csv
+head -n 10000 {snakemake.input.sheet}
 
 # -------------------------------------------------------------------------------------------------
 # Execute bcl2fastq v2
 
 bcl2fastq \
     --barcode-mismatches {barcode_mismatches} \
-    --sample-sheet {snakemake.params.output_dir}/SampleSheet.csv \
+    --sample-sheet {snakemake.input.sheet} \
     --runfolder-dir {snakemake.params.input_dir} \
     --output-dir $TMPDIR/demux_out \
     --interop-dir $TMPDIR/interop_dir \
     --processing-threads {bcl2fastq_threads} \
-    {bases_mask} {snakemake.params.tiles_arg}
+    --use-bases-mask {bases_mask_illumina} \
+    {snakemake.params.tiles_arg}
 
 tree $TMPDIR/demux_out
 
@@ -62,25 +72,45 @@ srcdir=$TMPDIR/demux_out/Project
 for path in $srcdir/*; do
     sample=$(basename $path | rev | cut -d _ -f 5- | rev)
     lane=$(basename $path | rev | cut -d _ -f 3 | rev)
-    dest={snakemake.params.output_dir}/$sample/$flowcell/$lane/$(basename $path)
-
+    dest={snakemake.params.output_dir}/$sample/$flowcell/$lane/{bases_mask}__$(basename $path)
+    mkdir -p $(dirname $dest)
     cp -dR $path $dest
-    pushd $(dirname $dest)
-    md5sum $(basename $dest) >$(basename $dest).md5
-    popd
 done
 
 # Move undetermined FASTQ files.
 srcdir=$TMPDIR/demux_out
 
 for path in $srcdir/Undetermined_*; do
-    lane=$(basename $path | rev | cut -d _ -f 3 | rev)
-    dest={snakemake.params.output_dir}/Undetermined/$flowcell/$lane/$(basename $path)
-
-    cp -dR $path $dest
-    pushd $(dirname $dest)
-    md5sum $(basename $dest) >$(basename $dest).md5
-    popd
+    if [[ -f $path ]]; then
+        lane=$(basename $path | rev | cut -d _ -f 3 | rev)
+        dest={snakemake.params.output_dir}/Undetermined/$flowcell/$lane/{bases_mask}__$(basename $path)
+        mkdir -p $(dirname $dest)
+        cp -dR $path $dest
+    fi
 done
+
+# Write out the html report files as an archive
+srcdir=$TMPDIR/demux_out
+pushd $srcdir
+tar -czvf {snakemake.params.output_dir}/html_report_{bases_mask}.tar.gz Reports
+popd
+
+touch {snakemake.output.marker}
 """
 )
+
+# bcl2fastq2 starts to count samples from 0 for each run. This breaks snakemake's assumption that
+# all samples from all sample sheets are numbered incrementally. Here, we look up the correct
+# sample number and replace it in the path. Easier to do in python than in bash above.
+fls = glob.glob(
+    os.path.join(snakemake.params.output_dir, "*/*/*/" + bases_mask + "__*.fastq.gz")  # noqa
+)
+for path in fls:
+    f = os.path.basename(path)
+    name = f.split("__")[1]  # remove $bases_mask__
+    name_elements = name.split("_")[:-3]
+    oldS = name_elements[-1]
+    name = "_".join(name_elements[:-1])
+    newS = sample_map[name]
+    newpath = path.replace("_".join([name, oldS]), "_".join([name, newS]))
+    os.rename(path, newpath)
